Adding indicators & context layer form of dashboard config

+++ django_project/frontend/src/pages/Admin/Components/List/index.jsx
@@ -26,44 +26,53 @@ import './style.scss';
 export function COLUMNS_ACTION(params, redirectUrl, editUrl = null, detailUrl = null) {
   editUrl = editUrl ? editUrl : urls.api.edit;
   detailUrl = detailUrl ? detailUrl : urls.api.detail;
-  return [
-    <GridActionsCellItem
-      icon={
-        <Tooltip title={`Edit ${params.row.name}`}>
-          <a
-            href={editUrl.replace('/0', `/${params.id}`)}>
-            <EditIcon/>
-          </a>
-        </Tooltip>
-      }
-      label="Edit"
-    />,
-    <GridActionsCellItem
-      className='AdminTableDelete'
-      icon={
-        <Tooltip title={`Delete ${params.row.name}`}>
-          <DeleteIcon/>
-        </Tooltip>
-      }
-      label="Delete"
-      onClick={
-        () => {
-          const api = detailUrl.replace('/0', `/${params.id}`);
-          if (confirm(`Are you sure you want to delete : ${params.row.name}?`)) {
-            $.ajax({
-              url: api,
-              method: 'DELETE',
-              success: function () {
-                window.location = redirectUrl;
-              },
-              beforeSend: beforeAjaxSend
-            });
-            return false;
+
+  const actions = []
+  if (editUrl) {
+    actions.push(
+      <GridActionsCellItem
+        icon={
+          <Tooltip title={`Edit ${params.row.name}`}>
+            <a
+              href={editUrl.replace('/0', `/${params.id}`)}>
+              <EditIcon/>
+            </a>
+          </Tooltip>
+        }
+        label="Edit"
+      />
+    )
+  }
+  if (detailUrl) {
+    actions.push(
+      <GridActionsCellItem
+        className='AdminTableDelete'
+        icon={
+          <Tooltip title={`Delete ${params.row.name}`}>
+            <DeleteIcon/>
+          </Tooltip>
+        }
+        label="Delete"
+        onClick={
+          () => {
+            const api = detailUrl.replace('/0', `/${params.id}`);
+            if (confirm(`Are you sure you want to delete : ${params.row.name}?`)) {
+              $.ajax({
+                url: api,
+                method: 'DELETE',
+                success: function () {
+                  window.location = redirectUrl;
+                },
+                beforeSend: beforeAjaxSend
+              });
+              return false;
+            }
           }
         }
-      }
-    />,
-  ]
+      />
+    )
+  }
+  return actions
 }
 
 /**
@@ -82,7 +91,7 @@ export function COLUMNS(pageName, redirectUrl, editUrl = null, detailUrl = null)
     { field: 'id', headerName: 'id', hide: true, width: 30 },
     { field: 'name', headerName: pageName + ' Name', flex: 1 },
     { field: 'description', headerName: 'Description', flex: 1 },
-    { field: 'group', headerName: 'Category', flex: 1 },
+    { field: 'category', headerName: 'Category', flex: 1 },
     {
       field: 'actions',
       type: 'actions',
@@ -99,20 +108,23 @@ export function COLUMNS(pageName, redirectUrl, editUrl = null, detailUrl = null)
  * @param {list} columns Columns setup.
  * @param {String} pageName Page Name.
  * @param {String} listUrl Url for list row.
+ * @param {list} initData Init Data.
  */
 export default function List(
   {
-    columns, pageName, listUrl
+    columns, pageName, listUrl, initData
   }
 ) {
-  const [data, setData] = useState(null);
+  const [data, setData] = useState(initData);
   const [search, setSearch] = useState(null);
 
   /** Fetch list of data */
   const fetchData = (url) => {
-    fetchingData(url, {}, {}, (data) => {
-      setData(data)
-    })
+    if (!data) {
+      fetchingData(url, {}, {}, (data) => {
+        setData(data)
+      })
+    }
   }
 
   // Show modal when url changed
@@ -130,7 +142,7 @@ export default function List(
   if (search) {
     rows = rows.filter(row => {
       return row.name.toLowerCase().includes(search)
-        || row.group.toLowerCase().includes(search)
+        || row.category.toLowerCase().includes(search)
         || row.description.toLowerCase().includes(search)
     })
   }
